RSS Aggregator
This project contains JAVA files and test cases for creating a list of
news stories from RSS feeds. The user may specify an RSS feed to pull
from and a keyword to search the feed for. The test cases test CNN and
Yahoo! news RSS feeds for stories with the keyword "Obama."